requirements.txt
alembic==1.13.2
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
Flask==2.3.2
Flask-Admin==1.6.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.20.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.6
pydantic==1.10.7
PyJWT==2.8.0
pytest==7.3.1
python-dotenv==1.0.0
pytz==2024.1
six==1.16.0
SQLAlchemy==1.4.36
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2
